Change how scss files are passed to sass
I used to pass the content of the files to sass on stdin, but it makes
resolving the @import a pain: since sass doesn't know where the file
actually is, it can't resolve the import relatively to the file
location. Giving sass include directories with -I is impractical because
the path changes depending on where sass is invoked from and I don't
want to include absolute paths.

My fix is to tell Hakyll to compile files is styles/ by passing their
path to sass, instead of feeding their content to stdin.